requirements.txt
flask
flask-sqlalchemy
langchain
openai
python-dotenv
flask-migrate
pyodbc